requirements.txt
spotipy==2.20.0
argparse==1.4.0
more_itertools
pyyaml
tqdm